Removed pedestal calculation from blr functions

+++ invisible_cities/sierpe/blr.pxd
@@ -35,7 +35,7 @@ cpdef deconvolve_signal(double [:] signal_daq,
                         double     thr_trigger     = *,
                         int accum_discharge_length = *)
 
-cpdef deconv_pmt(np.ndarray[np.int16_t, ndim=2] pmtrwf,
+cpdef deconv_pmt(np.ndarray[double, ndim=2] pmtrwf,
                  double [:]                     coeff_c,
                  double [:]                     coeff_blr,
                  list                           pmt_active             = *,
